+++ src/main/java/edu/grinnell/csc207/texteditor/GapBuffer.java
@@ -12,17 +12,6 @@ public class GapBuffer {
     private int cursor = 0;
     private int size = 0;
 
-
-    public String checkCursor(String message) {
-        if(cursor <= arrLength) { 
-            return ("length is greater than or equal to cursor in ");
-        }
-        if(cursor > 0) { 
-            return ("length is negative in ");
-        }
-        return null;
-    }
-
     /**
      * inserts character to string buffer
      * @param ch - character we want to add
@@ -41,9 +30,7 @@ public class GapBuffer {
             gapEnd = gapStart + (arrLength / 2);
         } 
         gapBuffer[cursor] = ch;
-        checkCursor("insert");
         cursor++;
-        checkCursor("insert");
         size++;
         gapStart++;
     }
@@ -52,11 +39,11 @@ public class GapBuffer {
      * deletes a character from the string buffer
      */
     public void delete() {
-        System.out.println("cursor " + cursor + " size " + size);
-        if(cursor > 0) {
-            cursor -= 1;
-            gapBuffer[cursor] = ' ';
-            //size -= 1; For some reason, this makes it not acctually insert the space so you cannot see the delete
+        if (cursor > 0 && size != 0) {
+            gapBuffer[cursor - 1] = '\0';
+            size--;
+            gapStart--;
+            cursor--;
         }
     }
 
@@ -75,6 +62,7 @@ public class GapBuffer {
         return gapEnd;
     }
 
+
     /**
      * moves the cursor position to the left
      */
@@ -82,13 +70,11 @@ public class GapBuffer {
         if (cursor != 0) {
             if (size != arrLength) {
                 gapBuffer[gapEnd - 1] = gapBuffer[cursor - 1];
-                gapBuffer[cursor] = '\0';
+                gapBuffer[cursor - 1] = '\0';
                 gapEnd--;
                 gapStart--;
             }
-            checkCursor("insert");
             cursor -= 1;
-            checkCursor("insert");
         }
     }
 
@@ -102,10 +88,8 @@ public class GapBuffer {
                 gapBuffer[cursor] = '\0';
                 gapEnd++;
                 gapStart++;
+                cursor += 1;
             }
-            checkCursor("insert");
-            cursor += 1;
-            checkCursor("insert");
         }
     }